% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyteplot.R
\name{analyteplot}
\alias{analyteplot}
\title{Make an analyte plot}
\usage{
analyteplot(data, analyte, Title, Ylab = "mg/L", Xlab, Ylimit)
}
\arguments{
\item{data}{desc}

\item{analyte}{desc}

\item{Title}{desc}

\item{Ylab}{desc}

\item{Xlab}{desc}

\item{Ylimit}{desc}
}
\description{
Make an analyte plot
}
