% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdeq_extract.R
\name{mdeq_extract}
\alias{mdeq_extract}
\title{Extract MDEQ data}
\usage{
mdeq_extract(mdeq_file, data_file, first_time = FALSE)
}
\arguments{
\item{mdeq_file}{A .xlsx file}

\item{data_file}{The title of the output file}

\item{first_time}{Logical. Is this the first time this function has run? Defaults to FALSE.}
}
\value{
A cleaned .csv sheet
}
\description{
Extract MDEQ data
}
